add all logic and testing misc-app

+++ staff/sergio-luis/node/misc-app/logic/add-contact.js
@@ -4,48 +4,54 @@ require('../utils/string')
 const Email = require('../utils/email')
 const uid = require('../utils/uid')
 require('../utils/json')
+require('../utils/date')
+const { find } = require('../data/findData')
 
 module.exports = (userId, contact, callback) => {
-    // TODO validate userId (string, not empty) and callback (function)
-    if (typeof contact !== 'object') throw new TypeError(`${contact} is not an object`)
-
-    // TODO make it so that at least should have the following fields: (name || suranme) && (email || phone)
-
     const { name, surname, email, phone, birthdate, country } = contact
 
-    if (name)
-        String.validate.notVoid(name)
+    String.validate.notVoid(userId)
 
-    if (surname)
-        String.validate.notVoid(surname)
+    if (typeof contact !== 'object') throw new TypeError(`${contact} is not an object`)
+    
+    if(!((name || surname) && (email || phone))) return callback(new Error('should have name or surname and email or phone'))
+
+    if (name) String.validate.notVoid(name)
+
+    if (surname) String.validate.notVoid(surname)
 
     if (email) {
         String.validate.notVoid(email)
         Email.validate(email)
     }
 
-    if (phone)
-        String.validate.notVoid(phone)
+    if (phone) String.validate.notVoid(phone)
 
     if (birthdate) {
         String.validate.notVoid(birthdate)
-        //Date.validate(birthdate) // TODO create this polyfill
+        _Date.validate(birthdate) // format - 30/12/2020
     }
 
-    if (country)
-        String.validate.notVoid(country)
+    if (country) String.validate.notVoid(country)
 
-    // TODO check userId is valid (i.e. the user exists), otherwise throw error
 
-    contact.user = userId
 
-    const id = uid()
-
-    const file = `${id}.json`
-
-    fs.writeFile(path.join(__dirname, '..', 'data', 'contacts', file), JSON.prettify(contact), error => {
+    find({id:userId}, 'users', (error, [user]) => {
         if (error) return callback(error)
 
-        callback(null, id)
-    })
+        if (!user) return callback(new Error(`user with ${userId} does not exist`))
+    
+        contact.user = userId
+        contact.contactId = uid()
+
+        const file = `${contact.contactId}.json`
+        
+    
+        fs.writeFile(path.join(__dirname, '..', 'data', 'contacts', file), JSON.prettify(contact), error => {
+            if (error) return callback(error)
+    
+            callback(null, contact.contactId)
+        })
+    
+    }) 
 }
